Refactor all TS interpreter code into one file
It was getting confusing tracing things through legacy_adapter and
whatnot. The reality is that this glue code is gonna stick around. We're
probably not changing the legacy format anytime soon.
