+++ libs/ballot-interpreter/src/interpret_and_save_files.test.ts
@@ -2,7 +2,7 @@ import { electionFamousNames2021Fixtures } from '@votingworks/fixtures';
 import { DEFAULT_MARK_THRESHOLDS, SheetOf } from '@votingworks/types';
 import { ALL_PRECINCTS_SELECTION } from '@votingworks/utils';
 import { loadImageData } from '@votingworks/image-utils';
-import { interpretSheetAndSaveImages } from './interpret_and_save_files';
+import { interpretSheetAndSaveImages } from './interpret';
 import { tmpDir } from '../test/helpers/tmp';
 
 test('interprets ballot images and saves images for storage', async () => {
